Clear compiler warnings in clang OS X
/Users/pnm/code/libbson/src/bson/bson-decimal128.c:102:11: warning: implicit conversion loses integer precision: 'uint64_t' (aka 'unsigned long long') to 'uint32_t' (aka 'unsigned int') [-Wshorten-64-to-32]
/Users/pnm/code/libbson/src/bson/bson-decimal128.c:681:19: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
/Users/pnm/code/libbson/src/bson/bson-decimal128.c:690:19: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
/Users/pnm/code/libbson/src/bson/bson-timegm.c:569:15: warning: implicit conversion loses integer precision: 'int64_t' (aka 'long long') to 'int_fast32_t' (aka 'int') [-Wshorten-64-to-32]
/Users/pnm/code/libbson/src/bson/bson.c:2147:4: warning: ambiguous expansion of macro 'va_copy' [-Wambiguous-macro]
/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/../lib/clang/8.0.0/include/stdarg.h:43:9: note: expanding this definition of 'va_copy'
/Users/pnm/code/libbson/src/bson/bson-compat.h:158:9: note: other definition of 'va_copy'

+++ src/bson/bson-timegm.c
@@ -558,7 +558,7 @@ time2sub (struct bson_tm *const tmp,
       return WRONG;
    if (normalize_overflow (&yourtm.tm_mday, &yourtm.tm_hour, HOURSPERDAY))
       return WRONG;
-   y = yourtm.tm_year;
+   y = (int_fast32_t) yourtm.tm_year;
    if (normalize_overflow32 (&y, &yourtm.tm_mon, MONSPERYEAR))
       return WRONG;
    /*
